+++ packages/contracts/test/global_hooks.ts
@@ -1,0 +1,3 @@
+after('generate coverage report', async () => {
+    await (global as any).__coverage_subprovider__.writeCoverageAsync();
+});