+++ packages/deployer/src/compiler.ts
@@ -1,3 +1,4 @@
+import { promisify } from '@0xproject/utils';
 import * as ethUtil from 'ethereumjs-util';
 import * as _ from 'lodash';
 import * as path from 'path';
@@ -186,9 +187,8 @@ export class Compiler {
         }
 
         const fullSolcVersion = binPaths[contractSpecificSourceData.solcVersion];
-        const solcBinPath = `./solc/solc_bin/${fullSolcVersion}`;
-        const solcBin = require(solcBinPath);
-        const solcInstance = solc.setupMethods(solcBin);
+        // TODO: Cache the compilers that are already loaded. Can come in handy on the planes/in places with bad connectivity
+        const solcInstance = await promisify<solc.SolcInstance>(solc.loadRemoteVersion)(fullSolcVersion);
 
         utils.consoleLog(`Compiling ${fileName}...`);
         const source = this._contractSources[fileName];
@@ -210,11 +210,14 @@ export class Compiler {
                 this._solcErrors.add(normalizedErrMsg);
             });
         }
-
         const contractName = path.basename(fileName, constants.SOLIDITY_FILE_EXTENSION);
         const contractIdentifier = `${fileName}:${contractName}`;
         const abi: Web3.ContractAbi = JSON.parse(compiled.contracts[contractIdentifier].interface);
-        const unlinked_binary = `0x${compiled.contracts[contractIdentifier].bytecode}`;
+        const bytecode = `0x${compiled.contracts[contractIdentifier].bytecode}`;
+        const runtimeBytecode = `0x${compiled.contracts[contractIdentifier].runtimeBytecode}`;
+        const sourceMap = compiled.contracts[contractIdentifier].srcmap;
+        const sourceMapRuntime = compiled.contracts[contractIdentifier].srcmapRuntime;
+        const sources = _.keys(compiled.sources);
         const updated_at = Date.now();
         const contractNetworkData: ContractNetworkData = {
             solc_version: contractSpecificSourceData.solcVersion,
@@ -222,8 +225,12 @@ export class Compiler {
             source_tree_hash: sourceTreeHash,
             optimizer_enabled: this._optimizerEnabled,
             abi,
-            unlinked_binary,
+            bytecode,
+            runtime_bytecode: runtimeBytecode,
             updated_at,
+            source_map: sourceMap,
+            source_map_runtime: sourceMapRuntime,
+            sources,
         };
 
         let newArtifact: ContractArtifact;