+++ packages/deployer/src/solc/bin_paths.ts
@@ -3,14 +3,14 @@ export interface BinaryPaths {
 }
 
 export const binPaths: BinaryPaths = {
-    '0.4.10': 'soljson-v0.4.10+commit.f0d539ae.js',
-    '0.4.11': 'soljson-v0.4.11+commit.68ef5810.js',
-    '0.4.12': 'soljson-v0.4.12+commit.194ff033.js',
-    '0.4.13': 'soljson-v0.4.13+commit.fb4cb1a.js',
-    '0.4.14': 'soljson-v0.4.14+commit.c2215d46.js',
-    '0.4.15': 'soljson-v0.4.15+commit.bbb8e64f.js',
-    '0.4.16': 'soljson-v0.4.16+commit.d7661dd9.js',
-    '0.4.17': 'soljson-v0.4.17+commit.bdeb9e52.js',
-    '0.4.18': 'soljson-v0.4.18+commit.9cf6e910.js',
-    '0.4.19': 'soljson-v0.4.19+commit.c4cbbb05.js',
+    '0.4.10': 'v0.4.10+commit.f0d539ae',
+    '0.4.11': 'v0.4.11+commit.68ef5810',
+    '0.4.12': 'v0.4.12+commit.194ff033',
+    '0.4.13': 'v0.4.13+commit.fb4cb1a',
+    '0.4.14': 'v0.4.14+commit.c2215d46',
+    '0.4.15': 'v0.4.15+commit.bbb8e64f',
+    '0.4.16': 'v0.4.16+commit.d7661dd9',
+    '0.4.17': 'v0.4.17+commit.bdeb9e52',
+    '0.4.18': 'v0.4.18+commit.9cf6e910',
+    '0.4.19': 'v0.4.19+commit.c4cbbb05',
 };