+++ packages/dev-utils/src/index.ts
@@ -1,4 +1,3 @@
-export { RPC } from './rpc';
 export { BlockchainLifecycle } from './blockchain_lifecycle';
 export { web3Factory } from './web3_factory';
 export { constants as devConstants } from './constants';