+++ packages/web3-wrapper/src/index.ts
@@ -15,6 +15,7 @@ interface RawLogEntry {
 }
 
 export class Web3Wrapper {
+    public isZeroExWeb3Wrapper = true;
     private _web3: Web3;
     private _defaults: Partial<TxData>;
     private _jsonRpcRequestId: number;
@@ -100,6 +101,20 @@ export class Web3Wrapper {
         const normalizedAddresses = _.map(addresses, address => address.toLowerCase());
         return normalizedAddresses;
     }
+    public async takeSnapshotAsync(): Promise<number> {
+        const snapshotId = Number(await this._sendRawPayloadAsync<string>({ method: 'evm_snapshot', params: [] }));
+        return snapshotId;
+    }
+    public async revertSnapshotAsync(snapshotId: number): Promise<boolean> {
+        const didRevert = await this._sendRawPayloadAsync<boolean>({ method: 'evm_revert', params: [snapshotId] });
+        return didRevert;
+    }
+    public async mineBlockAsync(): Promise<void> {
+        await this._sendRawPayloadAsync<string>({ method: 'evm_mine', params: [] });
+    }
+    public async increaseTimeAsync(timeDelta: number): Promise<void> {
+        await this._sendRawPayloadAsync<string>({ method: 'evm_increaseTime', params: [timeDelta] });
+    }
     public async getLogsAsync(filter: Web3.FilterObject): Promise<Web3.LogEntry[]> {
         let fromBlock = filter.fromBlock;
         if (_.isNumber(fromBlock)) {
@@ -140,7 +155,7 @@ export class Web3Wrapper {
         const txHash = await promisify<string>(this._web3.eth.sendTransaction)(txData);
         return txHash;
     }
-    private async _sendRawPayloadAsync<A>(payload: Web3.JSONRPCRequestPayload): Promise<A> {
+    private async _sendRawPayloadAsync<A>(payload: Partial<Web3.JSONRPCRequestPayload>): Promise<A> {
         const sendAsync = this._web3.currentProvider.sendAsync.bind(this._web3.currentProvider);
         const response = await promisify<Web3.JSONRPCResponsePayload>(sendAsync)(payload);
         const result = response.result;