+++ packages/deployer/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0 - _TBD, 2018_
 
     * Check dependencies when determining if contracts should be recompiled (#408).
+    * Export the `Compiler` (#TBD)
 
 ## v0.1.0 - _February 16, 2018_